ai project abdul hanan/requirements.txt
flask
PyPDF2
python-docx
scikit-learn